requirements.txt
discord.py
pywin32
watchdog